`TOKEN_MAP` code review fixes
