+++ src/components/token/index.d.ts
@@ -1,6 +1,13 @@
 import { FunctionComponent, HTMLAttributes } from 'react';
 import { CommonProps } from '../common';
 import { IconType } from '../icon';
+import { EuiTokenMapDisplayOptions } from './token_map';
+export {
+  EuiTokenMapDisplayOptions,
+  TokenShape,
+  TokenColor,
+  EuiTokenMapType,
+} from './token_map';
 
 declare module '@elastic/eui' {
   /**
@@ -8,32 +15,7 @@ declare module '@elastic/eui' {
    *
    * @see './token.js'
   */
-
   export type TokenSize = 's' | 'm' | 'l';
-
-  export type TokenColor =
-    | 'tokenTint01'
-    | 'tokenTint02'
-    | 'tokenTint03'
-    | 'tokenTint04'
-    | 'tokenTint05'
-    | 'tokenTint06'
-    | 'tokenTint07'
-    | 'tokenTint08'
-    | 'tokenTint09'
-    | 'tokenTint10'
-    | 'tokenTint11'
-    | 'tokenTint12';
-
-  export type TokenShape = 'circle' | 'square' | 'rectangle';
-
-  export interface EuiTokenMapDisplayOptions {
-    color?: TokenColor;
-    shape?: TokenShape;
-    fill?: boolean;
-    hasBorder?: boolean;
-  }
-
   export interface EuiTokenProps {
     iconType: IconType;
     size?: TokenSize;
